Update ideologies.js
Marxism-Leninism is very hard to get in this quiz. As a Marxist-Leninist myself, I didn't get it first try. So these are my proposed changes to make it more accurate. MLs are not inherently nationalistic as they believe in the eventual withering of the state. They also are not that conservative, just I lowered it by 10.